update to log file placement
