+++ backend/src/middleware/index.ts
@@ -1,2 +1,3 @@
 export { database } from "./mongoDB";
+export * as logs from "./logs";
 export * from "./errorHandler";
