+++ backend/src/utilities/index.ts
@@ -1,2 +1,1 @@
-export * as logs from "./logs";
 export { ErrorWithCode } from "./customError/ErrorWithCode";
